roundtable_dockerize/topic_api/topic/requirements.txt
click
h11
pydantic
starlette
fastapi
PyJWT
passlib
uvicorn
python-dotenv
sentence_transformers
scipy
spacy